Guarantee migration file deletion and use repo_path

+++ spec/tapioca/cli/dsl_spec.rb
@@ -449,27 +449,30 @@ module Tapioca
 
       it "aborts if there are pending migrations" do
         timestamp = Time.now.strftime("%Y%m%d%H%M")
-        migration_path = File.expand_path("../../support/repo/db/migrate/#{timestamp}_create_articles.rb", __dir__)
-        File.write(migration_path, <<~RUBY)
-          class CreateArticles < ActiveRecord::Migration[6.1]
-            def change
-              create_table(:articles) do |t|
-                t.timestamps
+        migration_path = repo_path / "db/migrate/#{timestamp}_create_articles.rb"
+
+        begin
+          File.write(migration_path, <<~RUBY)
+            class CreateArticles < ActiveRecord::Migration[6.1]
+              def change
+                create_table(:articles) do |t|
+                  t.timestamps
+                end
               end
             end
-          end
-        RUBY
+          RUBY
 
-        expected_output = <<~OUTPUT
-          Run `bin/rails db:migrate` to update your database then try again.
-          You have 1 pending migration:
-          #{timestamp}_create_articles.rb
-        OUTPUT
+          expected_output = <<~OUTPUT
+            Run `bin/rails db:migrate` to update your database then try again.
+            You have 1 pending migration:
+            #{timestamp}_create_articles.rb
+          OUTPUT
 
-        output = execute("dsl")
-        assert_match(expected_output, output)
-
-        File.delete(migration_path)
+          output = execute("dsl")
+          assert_match(expected_output, output)
+        ensure
+          File.delete(migration_path) if File.exist?(migration_path)
+        end
       end
 
       describe("verify") do
